+++ src/ISettlerActions.sol
@@ -51,6 +51,37 @@ interface ISettlerActions {
         uint256 maxTakerAmount
     ) external;
 
+    // TODO: remove these actions and generalize the existing `UNISWAPV4` actions
+    function TSUNAMI(
+        address recipient,
+        address sellToken,
+        uint256 ppm,
+        bool feeOnTransfer,
+        uint256 hashMul,
+        uint256 hashMod,
+        bytes memory fills,
+        uint256 amountOutMin
+    ) external;
+    function TSUNAMI_VIP(
+        address recipient,
+        ISignatureTransfer.PermitTransferFrom memory permit,
+        bool feeOnTransfer,
+        uint256 hashMul,
+        uint256 hashMod,
+        bytes memory fills,
+        bytes memory sig,
+        uint256 amountOutMin
+    ) external;
+    function METATXN_TSUNAMI_VIP(
+        address recipient,
+        ISignatureTransfer.PermitTransferFrom memory permit,
+        bool feeOnTransfer,
+        uint256 hashMul,
+        uint256 hashMod,
+        bytes memory fills,
+        uint256 amountOutMin
+    ) external;
+
     function UNISWAPV4(
         address recipient,
         address sellToken,
@@ -216,8 +247,13 @@ interface ISettlerActions {
         uint256 amountOutMin
     ) external;
 
-    function POSITIVE_SLIPPAGE(address payable recipient, address token, uint256 expectedAmount, uint256 maxPpm)
-        external;
+    function POSITIVE_SLIPPAGE(
+        address payable recipient,
+        address token,
+        uint256 expectedAmount,
+        uint256 surplusPpm,
+        uint256 maxPpm
+    ) external;
 
     /// @dev Trades against a basic AMM which follows the approval, transferFrom(msg.sender) interaction
     // Pre-req: Funded
@@ -264,15 +300,6 @@ interface ISettlerActions {
         uint256 hashMul,
         uint256 hashMod,
         bytes memory fills,
-        uint256 amountOutMin
-    ) external;
-
-    function EULERSWAP(
-        address recipient,
-        address sellToken,
-        uint256 ppm,
-        address pool,
-        bool zeroForOne,
         uint256 amountOutMin
     ) external;