+++ src/chains/Avalanche/Common.sol
@@ -7,7 +7,6 @@ import {IERC20} from "@forge-std/interfaces/IERC20.sol";
 import {DodoV2, IDodoV2} from "../../core/DodoV2.sol";
 import {UniswapV4} from "../../core/UniswapV4.sol";
 import {IPoolManager} from "../../core/UniswapV4Types.sol";
-import {EulerSwap, IEVC, IEulerSwap} from "../../core/EulerSwap.sol";
 import {BalancerV3} from "../../core/BalancerV3.sol";
 import {Bebop} from "../../core/Bebop.sol";
 import {FreeMemory} from "../../utils/FreeMemory.sol";
@@ -29,7 +28,7 @@ import {AVALANCHE_POOL_MANAGER} from "../../core/UniswapV4Addresses.sol";
 import {SettlerSwapAbstract} from "../../SettlerAbstract.sol";
 import {Permit2PaymentAbstract} from "../../core/Permit2PaymentAbstract.sol";
 
-abstract contract AvalancheMixin is FreeMemory, SettlerBase, DodoV2, UniswapV4, BalancerV3, EulerSwap, Bebop {
+abstract contract AvalancheMixin is FreeMemory, SettlerBase, DodoV2, UniswapV4, BalancerV3, Bebop {
     constructor() {
         assert(block.chainid == 43114 || block.chainid == 31337);
     }
@@ -56,11 +55,6 @@ abstract contract AvalancheMixin is FreeMemory, SettlerBase, DodoV2, UniswapV4, 
             ) = abi.decode(data, (address, IERC20, uint256, bool, uint256, uint256, bytes, uint256));
 
             sellToUniswapV4(recipient, sellToken, ppm, feeOnTransfer, hashMul, hashMod, fills, amountOutMin);
-        } else if (action == uint32(ISettlerActions.EULERSWAP.selector)) {
-            (address recipient, IERC20 sellToken, uint256 ppm, IEulerSwap pool, bool zeroForOne, uint256 amountOutMin) =
-                abi.decode(data, (address, IERC20, uint256, IEulerSwap, bool, uint256));
-
-            sellToEulerSwap(recipient, sellToken, ppm, pool, zeroForOne, amountOutMin);
         } else if (action == uint32(ISettlerActions.BALANCERV3.selector)) {
             (
                 address recipient,
@@ -118,10 +112,6 @@ abstract contract AvalancheMixin is FreeMemory, SettlerBase, DodoV2, UniswapV4, 
 
     function _POOL_MANAGER() internal pure override returns (IPoolManager) {
         return AVALANCHE_POOL_MANAGER;
-    }
-
-    function _EVC() internal pure override returns (IEVC) {
-        return IEVC(0xddcbe30A761Edd2e19bba930A977475265F36Fa1);
     }
 
     // I hate Solidity inheritance